requirements.txt
asgiref==3.11.1
Django==6.0.4
django-tailwind==4.4.2
djangorestframework==3.17.1
pillow==12.2.0
pytailwindcss==0.3.0
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
